feat: reduce sync memory usage
