+++ backend/app/core/config.py
@@ -1,14 +1,12 @@
 from pathlib import Path
 from typing import Literal
 
-from fastapi.templating import Jinja2Templates
 from pydantic import HttpUrl, computed_field, field_validator
 from pydantic_settings import BaseSettings
 
 
 class Settings(BaseSettings):
     PROJECT_NAME: str = "pypacktrends"
-    TEMPLATES: Jinja2Templates = Jinja2Templates(directory="app/templates")
     ENVIRONMENT: Literal["dev", "ci", "prod"] = "dev"
     SENTRY_DSN: HttpUrl | None = None
     SQLITE_DB_PATH: Path = (
@@ -35,4 +33,3 @@ class Settings(BaseSettings):
 
 
 settings = Settings()
-templates = settings.TEMPLATES
